Jamaican Bobsleigh Team in Sochi
It is official, the nation of Garvey, Marley, Bolt, Tessanne Chin and
the Cool Runnings phenomenon has done it again qualifying a team for the
2014 Olympic Winter Games in Sochi.

With scarce financial resources but powered by the fierce determination
of 46 year old team captain Winston Watt from Seven Roads in May Pen,
the Jamaican two-man team overcame all odds to qualify for the Olympics
against much better funded teams and much wealthier nations.  There is
currency and courage in the Jamaican spirit.

It has been a difficult road for Watt and his brakeman Marvin Dixon who
has sought to escape the troubled circumstances of his Rock Fort home;
and more work needs to be done.  The journey is not finished.

Existing obligations need to be paid, the costs of the training camp
before the Olympics need to be covered and Winston would like to be able
to purchase a set of jet fast runners like those used by the top team in
the world.

Let's help these brilliant Jamaicans.  Contribute today.  They have
taken many steps for their country.  Let us take a few steps for them.